+++ quiceh/Cargo.toml
@@ -77,7 +77,6 @@ smallvec = { version = "1.10", features = ["serde", "union"] }
 likely_stable = "0.1.2"
 cpu-time = "1.0.0"
 itertools = "0.12.1"
-pprof = { version = "0.14", features =  ["flamegraph", "criterion"] }
 
 [target."cfg(windows)".dependencies]
 winapi = { version = "0.3", features = ["wincrypt", "ws2def", "ws2ipdef", "ws2tcpip"] }
@@ -85,7 +84,8 @@ winapi = { version = "0.3", features = ["wincrypt", "ws2def", "ws2ipdef", "ws2tc
 [dev-dependencies]
 mio = { version = "0.8", features = ["net", "os-poll"] }
 url = "2.5"
-criterion = { version = "0.5.1", features = ["html_reports"] }
+criterion = { version = "0.5.1" }
+pprof = { version = "0.13", features =  ["flamegraph", "criterion"] }
 
 [lib]
 crate-type = ["lib", "staticlib", "cdylib"]
